Add wrapper for discrete actions
This wrapper serves as a way to get the old behaviour from
gym-trading-env back.

Note that we must in fact keep `position_index` when creating the
History object, because we cannot add any new column after that. We thus
set it to None by default in the env (which now uses continuous
actions), and set it to the correct value in the wrapper (which uses
discrete actions).