Fix integer check in Wrapper
We want to make sure that the `position_index` is an integer, because we
will use it to slice the positions list. But the NumPy int64, int32,
etc., can be seen as integers too, even though they do not inherint int.

This commit fixes the check to take into account NumPy values as well.