Fix bug with ndarray actions
Now that the action shape is a Box (instead of Discrete), when using
`.sample()`, it returns a ndarray, even when the Box's shape is `(1,)`.
That triggered a bug when taking an action, because the code silently
accepted the ndarray instead of an integer; this made the history
contain a ndarray of ndarray, which was problematic later when setting
some fields.

This commit adds a simple bugfix: if the action is a ndarray, we assume
it is of shape `(1,)`, and take its first (and only!) element instead.
We could add the assertion that the shape is `(1,)` to be really sure.