Add positional_range parameter
We previously set the action range to [-inf, +inf].
That avoided "forcing" a range for the user (they can have any position,
including very very very risky ones, such as -10_000 if they want so, or
even -1_000_000).
However, it had 2 very important drawbacks:
- When randomly sampling, the action was likely to immediately make the
  agent lose all its money. Because the range was infinite, the
  probability distribution function is likely to return an action that
  is so risky that the agent loses everything (e.g., -1_000_000).
- It was not easy to interpolate the result of a learning algorithm to a
  user-fixed range, such as [-1, +2] (or even [-10, +10]), because the
  initial range was infinite. In addition, clipping was likely to return
  always the same values (again, because the PDF is likely to return
  very high values, clipping to any "small" range would simply return
  one of the two bounds, e.g., either -10 or +10).

To avoid this, we now set the action range so that it avoids these
problems. To allow the user free choice, we let them set a custom value
in the Env constructor. By default, we use [-1, 2] which seems a
reasonable range: allow up to 100% borrowing.