Fix position in renders
The renders expect Python floats for the "position" column (i.e., what
the agent did). However, since the switch to a Box action space (rather
than a Discrete as previously), the position usually are NumPy ndarrays
of NumPy floats.

The ndarray was already converted to a single value in a previous commit;
but the value was still a NumPy float instead of a Python float. This is
changed in this commit; and we take into account the usual case of a
ndarray containing np floats.

We bump the version to 1.0.4